misc: Move shell_to_regex translation helper to filters where it's used
